requirements.txt
# Core Django Framework (keep versions pinned)
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.7

# Environment & Configuration
python-decouple==3.8
python-dotenv>=1.1.1
dj-database-url>=3.0.1

# Image Processing (let pip resolve versions)
Pillow>=10.0.1
cloudinary>=1.36.0
opencv-python>=4.6.0.66
opencv-contrib-python>=4.10.0.84

# Data Processing (let pip resolve numpy/pandas compatibility)
pandas>=2.1.4
numpy>=1.24.4,<2.0.0
openpyxl>=3.1.2
xlrd>=1.2.0
xlsxwriter>=3.2.5

# HTTP & API
requests>=2.31.0
requests-toolbelt>=1.0.0
aiohttp>=3.12.15
httpx>=0.28.1
httpx-sse>=0.4.1
httpcore>=1.0.9
h11>=0.16.0

# Async & Networking
aiohappyeyeballs>=2.6.1
aiosignal>=1.4.0
anyio>=4.10.0
frozenlist>=1.7.0
multidict>=6.6.4
yarl>=1.20.1

# GCP Services
google-cloud-storage>=2.10.0
google-cloud-vision>=3.4.0
django-storages>=1.14.2
google-api-core>=2.25.1
google-auth>=2.40.3
google-cloud-core>=2.4.3
google-crc32c>=1.7.1
google-resumable-media>=2.7.2
googleapis-common-protos>=1.70.0

# Additional Django Packages
django-cors-headers>=4.3.1
django-extensions>=3.2.3

# AI/ML Packages (let pip resolve versions)
langchain>=0.3.27
langchain-community>=0.3.27
langchain-core>=0.3.74
langchain-openai>=0.3.30
langchain-text-splitters>=0.3.9
langsmith>=0.4.14
openai>=1.100.1
tiktoken>=0.11.0
tokenizers>=0.21.4

# OCR Packages - Using DeepSeek API instead (via openai package)
# paddleocr>=3.2.0  # REMOVED - Using DeepSeek API
# paddlepaddle>=3.1.1  # REMOVED - Using DeepSeek API
# paddlex>=3.2.1  # REMOVED - Using DeepSeek API
# pytesseract>=0.3.13  # REMOVED - Not needed

# Background Tasks
celery>=5.5.3
redis>=6.4.0
amqp>=5.3.1
billiard>=4.2.1
kombu>=5.5.4
vine>=5.1.0

# Web Framework (Alternative)
fastapi>=0.116.1
uvicorn>=0.35.0
starlette>=0.47.3
python-multipart>=0.0.20

# Data Validation & Serialization
pydantic>=2.11.7
pydantic-core>=2.33.2
pydantic-settings>=2.10.1
marshmallow>=3.26.1
dataclasses-json>=0.6.7
orjson>=3.11.2
ujson>=5.10.0

# Machine Learning & Data Science (let pip resolve versions)
scikit-learn>=1.7.1
scipy>=1.16.1
joblib>=1.5.1
einops>=0.8.1
opt-einsum>=3.3.0
networkx>=3.5

# Utilities
click>=8.2.1
click-didyoumean>=0.3.1
click-plugins>=1.1.1.2
click-repl>=0.3.0
colorama>=0.4.6
colorlog>=6.9.0
tqdm>=4.67.1
prettytable>=3.16.0
tenacity>=9.1.2

# Parsing & Processing
beautifulsoup4>=4.13.4
lxml>=6.0.0
cssselect>=1.3.0
cssutils>=2.11.1
premailer>=3.10.0
pypdfium2>=4.30.0
pyclipper>=1.3.0.post6
shapely>=2.1.1
imagesize>=1.4.1

# Security & Cryptography
certifi>=2025.8.3
pycryptodome>=3.23.0
rsa>=4.9.1
pyasn1>=0.6.1
pyasn1-modules>=0.4.2

# System & OS
psutil>=7.0.0
GPUtil>=1.4.0
py-cpuinfo>=9.0.0
distro>=1.9.0
filelock>=3.19.1

# Cloud & Storage
fsspec>=2025.7.0
bce-python-sdk>=0.9.45

# Development Tools
ngrok>=1.4.0
prompt-toolkit>=3.0.52
wcwidth>=0.2.13

# Type Checking & Validation
typing-extensions>=4.14.1
typing-inspect>=0.9.0
typing-inspection>=0.4.1
annotated-types>=0.7.0
mypy-extensions>=1.1.0

# JSON & Data Formats
jsonpatch>=1.33
jsonpointer>=3.0.0
PyYAML>=6.0.2
ruamel.yaml>=0.18.14
ruamel.yaml.clib>=0.2.12

# Text Processing
regex>=2025.7.34
ftfy>=6.3.1
chardet>=5.2.0
charset-normalizer>=3.4.3
idna>=3.10

# Date & Time
python-dateutil>=2.9.0.post0
pytz>=2025.2
tzdata>=2025.2

# Protocol Buffers
protobuf>=6.32.0
proto-plus>=1.26.1

# Compression
zstandard>=0.24.0

# Other Dependencies (let pip resolve)
asgiref>=3.9.1
attrs>=25.3.0
cachetools>=5.5.2
common>=0.1.2
et-xmlfile>=2.0.0
future>=1.0.0
greenlet>=3.2.4
huggingface-hub>=0.34.4
Jinja2>=3.1.6
jiter>=0.10.0
MarkupSafe>=3.0.2
more-itertools>=10.7.0
packaging>=25.0
propcache>=0.3.2
setuptools>=65.5.0
six>=1.17.0
sniffio>=1.3.1
soupsieve>=2.7
SQLAlchemy>=2.0.43
sqlparse>=0.5.3
threadpoolctl>=3.6.0
urllib3>=2.5.0
aistudio-sdk>=0.3.6
modelscope>=1.29.2
